+++ jdaviz/configs/imviz/plugins/rotate_canvas/__init__.py
@@ -1,1 +1,0 @@
-from .rotate_canvas import *  # noqa